refactor(layout): isolate buffer view implementation
